requirements.txt
numpy>=1.21.0
torch>=1.9.0
transformers>=4.11.0
redis>=4.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
python-dotenv>=0.19.0
psycopg2-binary>=2.9.1
pytest>=6.2.5
black>=21.7b0
flake8>=3.9.0
gputil>=1.4.0
psutil>=5.8.0
nvidia-ml-py3>=7.352.0
websockets>=10.0
jinja2>=3.0.0
aiofiles>=0.7.0
python-multipart>=0.0.5
Pillow>=9.0.0
soundfile>=0.10.3
opencv-python>=4.5.3
torchaudio>=0.9.0
torchvision>=0.10.0
networkx>=2.6.0
qutip>=4.6.0
scipy>=1.7.0
matplotlib>=3.4.0
pandas>=1.3.0
scikit-learn>=0.24.0
triton>=3.0.0
safetensors>=0.3.0
einops>=0.6.0
accelerate>=0.21.0
bitsandbytes>=0.41.0
pyautogui>=0.9.53
keyboard>=0.13.5
mouse>=0.7.1
pywin32>=305
pytesseract>=0.3.8
tqdm>=4.62.3
flash-attn>=2.3.0
zstandard>=0.21.0
mmap-io>=0.1.0
rich>=10.0.0 